feat(Trame): Python Web App builder
Trame allow any Python savy to build a web interface to their
Python based application and create a rich and interactive application.

Trame also has a deep integration with VTK and ParaView to support
Remote and Local rendering of 3D dataset.

Many usage examples are available under ./examples/ for reference.

Documentation will soon be available on the Web Site of the project.