feat(setup_dev): add setup_dev() function for reloading trame applications
This function takes modules to reload when the reload button is pressed, which is visible if the
`--dev` argument is passed. This appears to be working both for running
`trame --dev <single_file>`, and by calling the `setup_dev()` function in a custom trame
application, along with passing the `--dev` flag to the python command. The function will
additionally call an `on_reload` function in any module it finds, and pass it another function that
can be used to reload more modules.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>