chore(docker): add container for cloud deployment
This allows trame apps to mount a special "/deploy" directory into the container that includes
instructions for how to install their trame app, and how to run it. This container will then run
a service that allows multiple clients to simultaneously connect to their own separate trame
applications at the web address for the server.

This also includes a script which fixes the "Docker and host filesystem owner matching problem".

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>