chore(docker): add conda docker image
This places common Dockerfile code in `Dockerfile.common`, and adds a `Dockerfile.pip` and
`Dockerfile.conda` which are based off of it. It also adds scripts unique to conda and pip in
`conda` and `pip` subdirectories, which get copied via their respective Dockerfiles.

This also adds conda builds to CI.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>