feat(paraview): add paraview to sys.path when requested
Trame can now find and add the paraview python library path to the sys.path if `trame.env.paraview`
was imported and the path to paraview was specified. It can be specified either via the
`TRAME_PARAVIEW` environment variable or a `--paraview <path>` argument on the command line. This
enables paraview to be imported and used.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>