trame a framework to write ubiquitous applications in Python
Version 2.0 is a rewrite of trame 1.x.
It keeps the idea behind trame while normalizing its API and enabling a wider extensability.
This new version open another set of doors while renforcing the strength of the trame core.

BREAKING CHANGE: The API between v1 and v2 are not compatible but the migration should remain straight forward

Co-authored-by: Patrick Avery <patrick.avery@kitware.com>