fix(docker): add `wheel` to pip docker image
Pip is now deprecating installation via the older install method that
does not use wheel. This deprecation message is now displaying, and it
appears that in a version of pip soon, we may start seeing errors.

Add wheel to the trame docker image to keep the installation process
simple for users (so that they will not be required to install wheel
themselves or create a `pyproject.toml` file).

See: https://github.com/pypa/pip/issues/8559

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>