fix(docker): add additional build options
This allows the user to separate out building the venv, launcher, and www.

The default behavior is that all will build if they do not exist. If they do exist,
then they will not be built.

However, the user can now select some options via a `TRAME_BUILD` environment variable.
If this string contains (single or any combination of) "venv", "launcher", or "www",
then those respective parts will be re-built, even if they already exist.

Additionally, a "no_www" can be specified in the `TRAME_BUILD` environment variable,
which indicates to skip the `www` generation, even if it does not exist.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>