fix(docker): move apps.json creation into launcher section
The `/deploy/setup` directory won't necessarily exist at runtime, because
it is optional if `/deploy/server` exists.

Thus, we should only create the `apps.json` file if we are performing one
of the other build steps (all of which require the `/deploy/server` directory).
It makes most sense to put this in the `launcher` section, since the launcher
uses this.

However, we must create the `apps.json` file before activating the venv,
because we have pyyaml in the root env, but not necessarily in the venv.
Also, the launcher creation script only uses built-in libraries, so it
can be moved before the venv as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>